Move ProductList classes to own package
