+++ app/src/main/java/kres/realtimeshoppinglist/firebase/productList/ProductListChangeListener.java
@@ -1,4 +1,4 @@
-package kres.realtimeshoppinglist.firebase;
+package kres.realtimeshoppinglist.firebase.productList;
 
 import android.util.Log;
